+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+- feat: add kubelet CPU manager settings
 - fix(rbac): allow pod deletion (after an eviction was attempted and failed)
 - feat: add support for EIP attachment to instances
 - fix(deps): update module github.com/exoscale/egoscale/v3 to v3.1.43 (#135)